coretelecoms/airflow/requirements.txt
apache-airflow-providers-amazon==8.25.0
apache-airflow-providers-google==10.20.0
apache-airflow-providers-postgres==5.11.0
pandas
pyarrow==15.0.0
boto3
psycopg2-binary
google-auth==2.29.0
google-auth-oauthlib==1.2.1
gspread==6.1.2
PyYAML==6.0.1
paramiko>=2.11
apache-airflow-providers-dbt-cloud
dbt-core
dbt-bigquery